refactor: extract the generate-quadlet command into its own module
The `scale` additions pushed main.rs past the 500-line limit; move
`write_quadlet` and its platform advisory (with their test) into a dedicated
`generate` module.
