+++ internal/main.rs
@@ -3,7 +3,6 @@
 // The binary carries no `unsafe`; deny it so any future addition is caught.
 #![deny(unsafe_code)]
 
-use std::path::Path;
 use std::process;
 
 #[cfg(feature = "completions")]
@@ -16,9 +15,11 @@ use tracing_subscriber::registry::LookupSpan;
 use tracing_subscriber::EnvFilter;
 
 mod cli;
+mod generate;
 mod resolve;
 
 use cli::*;
+use generate::write_quadlet;
 use resolve::*;
 
 /// Canonical project URL, reused for the bug-report hint on internal errors.
@@ -85,76 +86,6 @@ fn is_mutating(command: &Commands) -> bool {
 			| Commands::Restart { .. }
 			| Commands::Scale { .. }
 	)
-}
-
-/// Quadlet units are systemd unit files; they only run on Linux hosts (where
-/// systemd consumes them from `~/.config/containers/systemd/`). Generating them
-/// on macOS/Windows is legitimate (e.g. to deploy to a remote Linux host), so
-/// this returns an advisory string rather than blocking. `os` is
-/// [`std::env::consts::OS`]; the function is pure so every platform's branch is
-/// testable in a single run.
-fn quadlet_platform_advisory(os: &str) -> Option<String> {
-	(os != "linux").then(|| {
-		"quadlet units require systemd (Linux); generated files will not run on this host"
-			.to_string()
-	})
-}
-
-/// Generate Quadlet units from the compose file and either write them to a
-/// directory or print them to stdout. Warnings about unmapped fields go to
-/// stderr so stdout stays clean for piping.
-fn write_quadlet(
-	file: &podup::compose::types::ComposeFile,
-	project: &str,
-	output: Option<&Path>,
-) -> podup::Result<()> {
-	let result = podup::quadlet::generate(file, project);
-	if let Some(dup) = result.duplicate_filename() {
-		return Err(std::io::Error::new(
-			std::io::ErrorKind::InvalidInput,
-			format!(
-				"quadlet: two resources map to the same unit file {dup:?}; \
-				 rename one so their names do not collide after sanitization"
-			),
-		)
-		.into());
-	}
-	if let Some(advisory) = quadlet_platform_advisory(std::env::consts::OS) {
-		eprintln!("podup: warning: {advisory}");
-	}
-	for warning in &result.warnings {
-		eprintln!("podup: warning: {warning}");
-	}
-	match output {
-		Some(dir) => {
-			std::fs::create_dir_all(dir)?;
-			for unit in &result.units {
-				// Defense in depth: the unit stem is already sanitized in the
-				// library, but never write a unit whose name is anything but a
-				// plain file inside `dir` (rejects separators, `.` and `..`).
-				if Path::new(&unit.filename).file_name()
-					!= Some(std::ffi::OsStr::new(&unit.filename))
-				{
-					return Err(std::io::Error::new(
-						std::io::ErrorKind::InvalidInput,
-						format!("refusing unsafe quadlet unit file name: {}", unit.filename),
-					)
-					.into());
-				}
-				let path = dir.join(&unit.filename);
-				std::fs::write(&path, &unit.contents)?;
-				println!("wrote {}", path.display());
-			}
-		}
-		None => {
-			for unit in &result.units {
-				println!("# {}", unit.filename);
-				print!("{}", unit.contents);
-				println!();
-			}
-		}
-	}
-	Ok(())
 }
 
 #[tokio::main]
